refactor: move SeasonSelector from dashboard to archives
